C:\Users\korsaH\Documents\My Software Engineering Journey\Odin-recipe\Odin-recipe\css-exercises\animation\01-button-hover\style.css
It was very difficult for me in the beginning as to what to do. I googled out the solution but I could not find any because I wasn't specific. I naviagated further in the assignment and I found the solution. I copied it into the task and it worked. I am confident in going into my next task.